+++ tests/test_env_variables.py
@@ -16,9 +16,6 @@ CONTAINER_IMAGES = [LEAP_WITH_ENV]
 
 def test_environment_variables_present(auto_container: ContainerData):
     for env_var_name, env_var_val in ENV.items():
-        assert (
-            auto_container.connection.run_expect(
-                [0], f"echo ${env_var_name}"
-            ).stdout.strip()
-            == env_var_val
+        assert env_var_val == auto_container.remote.check_output(
+            f"echo ${env_var_name}"
         )