+++ tests/test_version.py
@@ -154,7 +154,7 @@ def test_podman_version_extract(stdout: str, ver: Version):
     assert _get_podman_version(stdout) == ver
 
 
-def test_container_runtime_parsing(host, container_runtime: OciRuntimeBase):
+def test_container_runtime_parsing(container_runtime: OciRuntimeBase):
     """Test that we can recreate the output of
     :command:`$container_runtime_binary --version` from the attribute
     :py:attr:`~pytest_container.runtime.OciRuntimeBase.version`.
@@ -165,18 +165,14 @@ def test_container_runtime_parsing(host, container_runtime: OciRuntimeBase):
         minor=container_runtime.version.minor,
         patch=container_runtime.version.patch,
     )
-    version_string = (
-        host.run_expect([0], f"{container_runtime.runner_binary} --version")
-        .stdout.strip()
-        .lower()
-    )
+    version_string = container_runtime.run_command("--version").lower()
 
     assert (
-        f"{container_runtime.runner_binary} version {version_without_build}"
+        f"{container_runtime.family} version {version_without_build}"
         in version_string
     )
 
-    if container_runtime.runner_binary == "docker":
+    if container_runtime.family == "docker":
         assert f"build {container_runtime.version.build}" in version_string