+++ tests/test_volumes.py
@@ -2,6 +2,7 @@
 import os
 from os.path import abspath
 from os.path import join
+from pathlib import Path
 from typing import List
 
 import pytest
@@ -13,7 +14,6 @@ from pytest_container.container import ContainerVolumeBase
 from pytest_container.container import DerivedContainer
 from pytest_container.container import get_volume_creator
 from pytest_container.container import VolumeFlag
-from pytest_container.runtime import LOCALHOST
 from pytest_container.runtime import OciRuntimeBase
 
 from .images import LEAP_URL
@@ -117,12 +117,10 @@ def test_container_host_volumes(container_per_test: ContainerData):
     for vol in container_per_test.container.volume_mounts:
         assert isinstance(vol, BindMount)
         assert vol.host_path
-        dir_on_host = LOCALHOST.file(vol.host_path)
-        assert dir_on_host.exists and dir_on_host.is_directory
+        dir_on_host = Path(vol.host_path)
+        assert dir_on_host.is_dir()
 
-        dir_in_container = container_per_test.connection.file(
-            vol.container_path
-        )
+        dir_in_container = container_per_test.remote.file(vol.container_path)
         assert dir_in_container.exists and dir_in_container.is_directory
 
 
@@ -141,10 +139,10 @@ def test_container_volume_host_writing(container_per_test: ContainerData):
     assert isinstance(vol, BindMount)
     assert vol.host_path
 
-    host_dir = LOCALHOST.file(vol.host_path)
-    assert not host_dir.listdir()
+    host_dir = Path(vol.host_path)
+    assert not list(host_dir.iterdir())
 
-    container_dir = container_per_test.connection.file(vol.container_path)
+    container_dir = container_per_test.remote.file(vol.container_path)
     assert not container_dir.listdir()
 
     contents = """This is just a testfile
@@ -154,7 +152,7 @@ def test_container_volume_host_writing(container_per_test: ContainerData):
     with open(join(vol.host_path, "test"), "w", encoding="utf8") as testfile:
         testfile.write(contents)
 
-    testfile_in_container = container_per_test.connection.file(
+    testfile_in_container = container_per_test.remote.file(
         join(vol.container_path, "test")
     )
     assert (
@@ -165,9 +163,7 @@ def test_container_volume_host_writing(container_per_test: ContainerData):
 
     # check that the file does not appear in the second mount
     vol2 = container_per_test.container.volume_mounts[1]
-    assert not container_per_test.connection.file(
-        vol2.container_path
-    ).listdir()
+    assert not container_per_test.remote.file(vol2.container_path).listdir()
 
 
 LEAP_WITH_CONTAINER_VOLUMES = DerivedContainer(
@@ -182,7 +178,7 @@ LEAP_WITH_CONTAINER_VOLUMES = DerivedContainer(
 def test_container_volumes(container_per_test: ContainerData):
     assert len(container_per_test.container.volume_mounts) == 2
     for vol in container_per_test.container.volume_mounts:
-        dir_in_container = container_per_test.connection.file(
+        dir_in_container = container_per_test.remote.file(
             join("/", vol.container_path)
         )
         assert dir_in_container.exists and dir_in_container.is_directory
@@ -195,15 +191,15 @@ def test_container_volume_writeable(container_per_test: ContainerData):
     vol = container_per_test.container.volume_mounts[0]
     assert isinstance(vol, ContainerVolume) and vol.volume_id
 
-    container_dir = container_per_test.connection.file(vol.container_path)
+    container_dir = container_per_test.remote.file(vol.container_path)
     assert not container_dir.listdir()
 
     contents = "OK"
-    container_per_test.connection.run_expect(
-        [0], f"bash -c 'echo -n {contents} > {vol.container_path}/test'"
+    container_per_test.remote.check_output(
+        f"bash -c 'echo -n {contents} > {vol.container_path}/test'"
     )
 
-    testfile_in_container = container_per_test.connection.file(
+    testfile_in_container = container_per_test.remote.file(
         join(vol.container_path, "test")
     )
     assert (
@@ -214,9 +210,7 @@ def test_container_volume_writeable(container_per_test: ContainerData):
 
     # check that the file does not appear in the second mount
     vol2 = container_per_test.container.volume_mounts[1]
-    assert not container_per_test.connection.file(
-        vol2.container_path
-    ).listdir()
+    assert not container_per_test.remote.file(vol2.container_path).listdir()
 
 
 LEAP_WITH_BIND_MOUNT_AND_VOLUME = DerivedContainer(
@@ -237,13 +231,11 @@ def test_concurrent_container_volumes(container_per_test: ContainerData):
     assert container_per_test.container.volume_mounts
 
     for vol in container_per_test.container.volume_mounts:
-        assert container_per_test.connection.file(vol.container_path).exists
-        assert not container_per_test.connection.file(
-            vol.container_path
-        ).listdir()
+        assert container_per_test.remote.file(vol.container_path).exists
+        assert not container_per_test.remote.file(vol.container_path).listdir()
 
-        container_per_test.connection.run_expect(
-            [0], f"echo > {vol.container_path}/test_file"
+        container_per_test.remote.check_output(
+            f"echo > {vol.container_path}/test_file"
         )
 
 
@@ -264,9 +256,14 @@ LEAP_WITH_ROOTDIR_BIND_MOUNTED = DerivedContainer(
 def test_bind_mount_cwd(container: ContainerData):
     vol = container.container.volume_mounts[0]
     assert isinstance(vol, BindMount)
-    assert container.connection.file("/src/").exists and sorted(
-        container.connection.file("/src/").listdir()
-    ) == sorted(LOCALHOST.file(vol.host_path).listdir())
+    assert vol.host_path is not None
+
+    assert container.remote.file("/src/").exists and sorted(
+        container.remote.file("/src/").listdir()
+    ) == sorted(
+        str(p.relative_to(vol.host_path))
+        for p in Path(vol.host_path).iterdir()
+    )
 
 
 def test_bind_mount_fails_when_host_path_not_present() -> None: